requirements.txt
alembic==0.9.9
click==6.7
Flask==1.0.2
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
google-api-python-client==1.6.7
gunicorn==19.8.1
httplib2==0.11.3
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
oauth2client==4.1.2
psycopg2-binary==2.7.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
python-dateutil==2.7.3
python-editor==1.0.3
rsa==3.4.2
six==1.11.0
SQLAlchemy==1.2.7
uritemplate==3.0.0
Werkzeug==0.14.1
